Anket içerisine anket yayındaysa anket içerisindeki sorularda düzenleme yapamaması için kod düzenlemesi yapıldı.

+++ ISUAnket.Business/Managers/SoruManager.cs
@@ -14,10 +14,12 @@ namespace ISUAnket.Business.Managers
     public class SoruManager : ISoruService
     {
         private readonly ISoruRepository _soruRepository;
+        private readonly IAnketRepository _anketRepository;
 
-        public SoruManager(ISoruRepository soruRepository)
+        public SoruManager(ISoruRepository soruRepository, IAnketRepository anketRepository)
         {
             _soruRepository = soruRepository;
+            _anketRepository = anketRepository;
         }
 
         public async Task<List<Soru>> GetListAllServiceAsync()
@@ -49,11 +51,41 @@ namespace ISUAnket.Business.Managers
 
         public async Task AddServiceAsync(Soru entity)
         {
+            #region Ankete yayınlanma durumuna göre soru eklememe kontrolü
+
+            if (entity.AnketId==null)
+            {
+                throw new ArgumentException(nameof(entity.AnketId), "Anket bilgisi boş olamaz!");
+            }
+            var anket=await _anketRepository.GetByIdAsync(entity.AnketId.Value);
+            if (anket?.AnketDurumu==EntityLayer.Enums.AnketDurumuEnum.Yayınlandı)
+            {
+                throw new InvalidOperationException("Yayınlanmış ankete soru eklenemez!");
+            }
+
+            #endregion
+
             await _soruRepository.AddAsync(entity);
         }
 
         public async Task UpdateServiceAsync(Soru entity)
         {
+            #region Yayınlanmış ankete soru ekleyemem kontrolü
+
+            if (entity.AnketId==null)
+            {
+                throw new ArgumentNullException(nameof(entity.AnketId), "Anket bilgisi boş olamaz!");
+            }
+
+            var anket = await _anketRepository.GetByIdAsync(entity.AnketId.Value);
+
+            if (anket?.AnketDurumu==EntityLayer.Enums.AnketDurumuEnum.Yayınlandı)
+            {
+                throw new InvalidOperationException("Yayınlanmış anketteki soru güncellenemez!");
+            }
+
+            #endregion
+
             await _soruRepository.UpdateAsync(entity);
         }
 
@@ -61,6 +93,22 @@ namespace ISUAnket.Business.Managers
 
         public async Task DeleteServiceAsync(Soru entity)
         {
+            #region Yayınlanmış ankete soru silememe kontrolü
+
+            if (entity.AnketId==null)
+            {
+                throw new ArgumentNullException(nameof(entity.AnketId), "Anket numarası boş olamaz!");
+            }
+
+            var anket = await _anketRepository.GetByIdAsync(entity.AnketId.Value);
+
+            if (anket?.AnketDurumu==EntityLayer.Enums.AnketDurumuEnum.Yayınlandı)
+            {
+                throw new InvalidOperationException("Yayınlanmış anketteki soru silinemez!");
+            }
+
+            #endregion
+
             await _soruRepository.DeleteAsync(entity);
         }
